+++ Simulator/Simulator/src/data_model/result.cpp
@@ -1,7 +1,7 @@
-#include "../../include/stdafx.h"
+
 #include "../../include/result.h"
 
-Result::Result(Instruction instruction, uint64_t result) : instruction(instruction), result(result)
+Result::Result(Instruction instruction, int result) : instruction(instruction), result(result)
 {
 }
